Revert io split
